# internal/api/gen/http/openapi3.yaml
openapi: 3.0.3
info:
  title: Enduro API
  version: "1.0"
servers:
- url: http://localhost:9000
paths:
...
  /collection/{id}/download:
    get:
      tags:
      - collection
      summary: download collection
      description: Download collection by ID
      operationId: collection#download
      parameters:
      - name: id
        in: path
        description: Identifier of collection to look up
        required: true
        schema:
          type: integer
          description: Identifier of collection to look up
          example: 12098715864266581887
        example: 13843919557982438648
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: string
                example:
                - 83
                - 117
                - 115
                - 99
                - 105
                - 112
                - 105
                - 116
                - 32
                - 105
                - 108
                - 108
                - 111
                - 32
                - 99
                - 117
                - 109
                - 113
                - 117
                - 101
                - 32
                - 115
                - 105
                - 116
                - 46
                format: binary
              example:
              - 72
              - 105
              - 99
              - 32
              - 100
              - 101
              - 98
              - 105
              - 116
              - 105
              - 115
              - 32
              - 97
              - 100
              - 32
              - 97
              - 110
              - 105
              - 109
              - 105
              - 32
              - 113
              - 117
              - 105
              - 97
              - 32
              - 117
              - 116
              - 32
              - 99
              - 117
              - 108
              - 112
              - 97
              - 46
        "404":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
              example:
                id: 18157397634821487954
                message: Itaque soluta sed et.
...
  /collection/{id}/workflow:
    get:
      tags:
      - collection
      summary: workflow collection
      description: Retrieve workflow status by ID
      operationId: collection#workflow
      parameters:
      - name: id
        in: path
        description: Identifier of collection to look up
        required: true
        schema:
          type: integer
          description: Identifier of collection to look up
          example: 10185585136924236988
        example: 7351783235052544585
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnduroCollectionWorkflowStatus'
              example:
                history:
                - details: Odio eos laboriosam rerum amet.
                  id: 3305209720667027557
                  type: Sapiente ea aut dicta.
                - details: Odio eos laboriosam rerum amet.
                  id: 3305209720667027557
                  type: Sapiente ea aut dicta.
                - details: Odio eos laboriosam rerum amet.
                  id: 3305209720667027557
                  type: Sapiente ea aut dicta.
                status: Ipsam quia eius vitae praesentium commodi voluptas.
        "404":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
              example:
                id: 9569705239864435356
                message: Aut sint.
...
components:
  schemas:
...
    EnduroCollectionWorkflowHistory:
      type: object
      properties:
        details:
          type: string
          description: Contents of the event
          example: Libero quasi et similique est.
          format: binary
        id:
          type: integer
          description: Identifier of collection
          example: 16019650542386041247
        type:
          type: string
          description: Type of the event
          example: Qui assumenda suscipit sit.
      description: WorkflowHistoryEvent describes a history event in Cadence.
      example:
        details: Dolores ea voluptatum unde laborum temporibus voluptas.
        id: 12804837024206468649
        type: Consequuntur est autem quam.
    EnduroCollectionWorkflowHistoryCollection:
      type: array
      items:
        $ref: '#/components/schemas/EnduroCollectionWorkflowHistory'
      example:
      - details: Odio eos laboriosam rerum amet.
        id: 3305209720667027557
        type: Sapiente ea aut dicta.
      - details: Odio eos laboriosam rerum amet.
        id: 3305209720667027557
        type: Sapiente ea aut dicta.
    EnduroCollectionWorkflowStatus:
      type: object
      properties:
        history:
          $ref: '#/components/schemas/EnduroCollectionWorkflowHistoryCollection'
        status:
          type: string
          example: Alias aut dolore.
      example:
        history:
        - details: Odio eos laboriosam rerum amet.
          id: 3305209720667027557
          type: Sapiente ea aut dicta.
        - details: Odio eos laboriosam rerum amet.
          id: 3305209720667027557
          type: Sapiente ea aut dicta.
        - details: Odio eos laboriosam rerum amet.
          id: 3305209720667027557
          type: Sapiente ea aut dicta.
        status: Sint voluptate consequatur.
...
    NotFound:
      type: object
      properties:
        id:
          type: integer
          description: Identifier of missing collection
          example: 2135324167707328023
        message:
          type: string
          description: Message of error
          example: Et autem adipisci et occaecati quaerat.
      description: Collection not found
      example:
        id: 11276450690720251243
        message: Ab pariatur.
      required:
      - message
      - id
...